requirements.txt
fastapi
uvicorn
paho-mqtt
python-dotenv
apscheduler
